pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "1.0.0"
description = ""
authors = [
	"ByMystick <16-@bk.ru>"
]
readme = "README.md"
packages = [
	{include = "gendiff"},
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^5.4.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.0"
pytest-cov = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
